[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14224 KB, Score: 100 point -EuiseonHub

+++ algorithm/baekjoon/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,8 +7,8 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Integer.parseInt(st.nextToken());
+        long b = Integer.parseInt(st.nextToken());
         System.out.print((a + b) * (a - b));
     }
 }
